+++ sbws/lib/v3bwfile.py
@@ -8,7 +8,9 @@ import os
 from statistics import median, mean
 
 from sbws import __version__
-from sbws.globals import SPEC_VERSION, BW_LINE_SIZE, SCALE_CONSTANT
+from sbws.globals import (SPEC_VERSION, BW_LINE_SIZE, SBWS_SCALE_CONSTANT,
+                          SBWS_SCALING, TORFLOW_BW_MARGIN, TORFLOW_SCALING,
+                          TORFLOW_OBS_LAST, TORFLOW_OBS_MEAN)
 from sbws.lib.resultdump import ResultSuccess, _ResultType
 from sbws.util.filelock import DirectoryLock
 from sbws.util.timestamp import now_isodt_str, unixts_to_isodt_str
@@ -38,7 +40,8 @@ BW_KEYVALUES_BASIC = ['node_id', 'bw']
 BW_KEYVALUES_FILE = BW_KEYVALUES_BASIC + \
                     ['master_key_ed25519', 'nick', 'rtt', 'time',
                      'success', 'error_stream', 'error_circ', 'error_misc']
-BW_KEYVALUES_EXTRA_BWS = ['bw_bs_median', 'bw_bs_mean', 'desc_avg_bw_bs']
+BW_KEYVALUES_EXTRA_BWS = ['bw_bs_median', 'bw_bs_mean', 'desc_avg_bw_bs',
+                          'desc_obs_bw_bs_last', 'desc_obs_bw_bs_mean']
 BW_KEYVALUES_EXTRA = BW_KEYVALUES_FILE + BW_KEYVALUES_EXTRA_BWS
 BW_KEYVALUES_INT = ['bw', 'rtt', 'success', 'error_stream',
                     'error_circ', 'error_misc'] + BW_KEYVALUES_EXTRA_BWS
@@ -260,6 +263,10 @@ class V3BWLine(object):
             kwargs['bw_bs_mean'] = cls.bw_bs_mean_from_results(success_results)
             kwargs['bw_bs_median'] = cls.bw_bs_median_from_results(
                 success_results)
+            kwargs['desc_obs_bw_bs_last'] = \
+                cls.desc_obs_bw_bs_last_from_results(success_results)
+            kwargs['desc_obs_bw_bs_mean'] = \
+                cls.desc_obs_bw_bs_mean_from_results(success_results)
             bwl = cls(node_id, bw, **kwargs)
             return bwl
         return None
@@ -313,6 +320,24 @@ class V3BWLine(object):
                         for rt in _ResultType])
         return rt_dict
 
+    @staticmethod
+    def desc_obs_bw_bs_mean_from_results(results):
+        desc_obs_bws = []
+        for r in results:
+            if r.relay_observed_bandwidth is not None:
+                desc_obs_bws.append(r.relay_observed_bandwidth)
+        if desc_obs_bws:
+            return max(round(mean(desc_obs_bws)), 1)
+        return None
+
+    @staticmethod
+    def desc_obs_bw_bs_last_from_results(results):
+        # the last is at the end of the list
+        for r in reversed(results):
+            if r.relay_observed_bandwidth is not None:
+                return r.relay_observed_bandwidth
+        return None
+
     @property
     def bw_keyvalue_tuple_ls(self):
         """Return list of KeyValue Bandwidth Line tuples."""
@@ -355,20 +380,52 @@ class V3BWFile(object):
         self.bw_lines = v3bwlines
 
     def __str__(self):
-        return str(self.header) + ''.join([str(bw_line)
+        return str(self.header) + ''.join([str(bw_line) or ''
                                            for bw_line in self.bw_lines])
 
     @classmethod
     def from_results(cls, results, state_fpath='',
-                     scale_constant=None):
-        bw_lines = [V3BWLine.from_results(results[fp]) for fp in results]
-        bw_lines = sorted(bw_lines, key=lambda d: d.bw, reverse=True)
-        if scale_constant is not None:
-            bw_lines = cls.bw_sbws_scale(bw_lines, scale_constant)
-            cls.warn_if_not_accurate_enough(bw_lines, scale_constant)
-        else:
-            bw_lines = cls.bw_kb(bw_lines)
+                     scale_constant=SBWS_SCALE_CONSTANT,
+                     scaling_method=None, torflow_obs=TORFLOW_OBS_LAST,
+                     reverse=False):
+        """Create V3BWFile class from sbws Results.
+
+        :param dict results: see below
+        :param str state_fpath: path to the state file
+        :param int scaling_method:
+            Scaling method to obtain the bandwidth
+            Posiable values: {NONE, SBWS_SCALING, TORFLOW_SCALING} = {0, 1, 2}
+        :param int scale_constant: sbws scaling constant
+        :param int torflow_obs: method to choose descriptor observed bandwidth
+        :param bool reverse: whether to sort the bw lines descending or not
+
+        Results are in the form::
+
+            {'relay_fp1': [Result1, Result2, ...],
+             'relay_fp2': [Result1, Result2, ...]}
+
+        """
+        # TODO: change scaling_method to TORFLOW_SCALING before getting this
+        # in production
+        log.info('Processing results to generate a bandwidth list file.')
         header = V3BWHeader.from_results(results, state_fpath)
+        bw_lines_raw = []
+        for fp in results.keys():
+            line = V3BWLine.from_results(results[fp])
+            if line is not None:
+                bw_lines_raw.append(line)
+        if not bw_lines_raw:
+            return cls(header, [])
+        if scaling_method == SBWS_SCALING:
+            bw_lines = cls.bw_sbws_scale(bw_lines_raw, scale_constant)
+            cls.warn_if_not_accurate_enough(bw_lines, scale_constant)
+            # log.debug(bw_lines[-1])
+        elif scaling_method == TORFLOW_SCALING:
+            bw_lines = cls.bw_torflow_scale(bw_lines_raw, torflow_obs)
+            # log.debug(bw_lines[-1])
+        else:
+            bw_lines = cls.bw_kb(bw_lines_raw)
+            # log.debug(bw_lines[-1])
         f = cls(header, bw_lines)
         return f
 
@@ -380,7 +437,7 @@ class V3BWFile(object):
         return sorted(bw_lines_scaled, key=lambda x: x.bw, reverse=reverse)
 
     @staticmethod
-    def bw_sbws_scale(bw_lines, scale_constant=SCALE_CONSTANT,
+    def bw_sbws_scale(bw_lines, scale_constant=SBWS_SCALE_CONSTANT,
                       reverse=False):
         """Return a new V3BwLine list scaled using sbws method.
 
@@ -417,7 +474,7 @@ class V3BWFile(object):
 
     @staticmethod
     def warn_if_not_accurate_enough(bw_lines,
-                                    scale_constant=SCALE_CONSTANT):
+                                    scale_constant=SBWS_SCALE_CONSTANT):
         margin = 0.001
         accuracy_ratio = median([l.bw for l in bw_lines]) / scale_constant
         log.info('The generated lines are within {:.5}% of what they should '
@@ -425,6 +482,198 @@ class V3BWFile(object):
         if accuracy_ratio < 1 - margin or accuracy_ratio > 1 + margin:
             log.warning('There was %f%% error and only +/- %f%% is '
                         'allowed', (1 - accuracy_ratio) * 100, margin * 100)
+
+    @staticmethod
+    def bw_torflow_scale(bw_lines, desc_obs_bws=TORFLOW_OBS_LAST,
+                         cap=TORFLOW_BW_MARGIN, reverse=False):
+        """
+        Obtain final bandwidth measurements applying Torflow's scaling
+        method.
+
+        From Torflow's README.spec.txt (section 2.2)::
+
+            In this way, the resulting network status consensus bandwidth values  # NOQA
+            are effectively re-weighted proportional to how much faster the node  # NOQA
+            was as compared to the rest of the network.
+
+        The variables and steps used in Torflow:
+
+        **strm_bw**::
+
+            The strm_bw field is the average (mean) of all the streams for the relay  # NOQA
+            identified by the fingerprint field.
+            strm_bw = sum(bw stream x)/|n stream|
+
+        **filt_bw**::
+
+            The filt_bw field is computed similarly, but only the streams equal to  # NOQA
+            or greater than the strm_bw are counted in order to filter very slow  # NOQA
+            streams due to slow node pairings.
+
+        **filt_sbw and strm_sbw**::
+
+            for rs in RouterStats.query.filter(stats_clause).\
+                  options(eagerload_all('router.streams.circuit.routers')).all():  # NOQA
+              tot_sbw = 0
+              sbw_cnt = 0
+              for s in rs.router.streams:
+                if isinstance(s, ClosedStream):
+                  skip = False
+                  #for br in badrouters:
+                  #  if br != rs:
+                  #    if br.router in s.circuit.routers:
+                  #      skip = True
+                  if not skip:
+                    # Throw out outliers < mean
+                    # (too much variance for stddev to filter much)
+                    if rs.strm_closed == 1 or s.bandwidth() >= rs.sbw:
+                      tot_sbw += s.bandwidth()
+                      sbw_cnt += 1
+
+            if sbw_cnt: rs.filt_sbw = tot_sbw/sbw_cnt
+            else: rs.filt_sbw = None
+
+        **filt_avg, and strm_avg**::
+
+            Once we have determined the most recent measurements for each node, we  # NOQA
+            compute an average of the filt_bw fields over all nodes we have measured.  # NOQA
+
+        ::
+
+            filt_avg = sum(map(lambda n: n.filt_bw, nodes.itervalues()))/float(len(nodes))  # NOQA
+            strm_avg = sum(map(lambda n: n.strm_bw, nodes.itervalues()))/float(len(nodes))  # NOQA
+
+        **true_filt_avg and true_strm_avg**::
+
+            for cl in ["Guard+Exit", "Guard", "Exit", "Middle"]:
+                true_filt_avg[cl] = filt_avg
+                true_strm_avg[cl] = strm_avg
+
+        In the non-pid case, all types of nodes get the same avg
+
+        **n.fbw_ratio and n.fsw_ratio**::
+
+            for n in nodes.itervalues():
+                n.fbw_ratio = n.filt_bw/true_filt_avg[n.node_class()]
+                n.sbw_ratio = n.strm_bw/true_strm_avg[n.node_class()]
+
+        **n.ratio**::
+
+            These averages are used to produce ratios for each node by dividing the  # NOQA
+            measured value for that node by the network average.
+
+        ::
+
+            # Choose the larger between sbw and fbw
+              if n.sbw_ratio > n.fbw_ratio:
+                n.ratio = n.sbw_ratio
+              else:
+                n.ratio = n.fbw_ratio
+
+        **desc_bw**:
+
+        It is the ``observed bandwidth`` in the descriptor, NOT the ``average
+        bandwidth``::
+
+            return Router(ns.idhex, ns.nickname, bw_observed, dead, exitpolicy,
+            ns.flags, ip, version, os, uptime, published, contact, rate_limited,  # NOQA
+            ns.orhash, ns.bandwidth, extra_info_digest, ns.unmeasured)
+            self.desc_bw = max(bw,1) # Avoid div by 0
+
+        **new_bw**::
+
+            These ratios are then multiplied by the most recent observed descriptor  # NOQA
+            bandwidth we have available for each node, to produce a new value for  # NOQA
+            the network status consensus process.
+
+        ::
+
+            n.new_bw = n.desc_bw*n.ratio
+
+        The descriptor observed bandwidth is multiplied by the ratio.
+        With empirical results this ratio is ~[0.9, 8.9]
+
+        **Limit the bandwidth to a maximum**::
+
+            NODE_CAP = 0.05
+
+        ::
+
+            if n.new_bw > tot_net_bw*NODE_CAP:
+              plog("INFO", "Clipping extremely fast "+n.node_class()+" node "+n.idhex+"="+n.nick+  # NOQA
+                   " at "+str(100*NODE_CAP)+"% of network capacity ("+
+                   str(n.new_bw)+"->"+str(int(tot_net_bw*NODE_CAP))+") "+
+                   " pid_error="+str(n.pid_error)+
+                   " pid_error_sum="+str(n.pid_error_sum))
+              n.new_bw = int(tot_net_bw*NODE_CAP)
+
+        However, tot_net_bw does not seems to be updated when not using pid.
+        This clipping would make faster relays to all have the same value.
+
+        All of that can be expressed as:
+
+        .. math::
+
+                bwnew_i &=
+                    max\\left(
+                        \\frac{bw_i}{\\mu},
+                        min \\left(
+                            bw_i,
+                            bw_i \\times \\mu
+                            \\right)
+                                \\times
+                                \\frac{bw}{\\sum_{i=1}^{n}
+                                min \\left(bw_i,
+                                    bw_i \\times \\mu
+                            \\right)}
+                        \\right)
+                    \\times bwdescobs_i \\
+
+                &=
+                    max\\left(
+                        \\frac{bw_i}{\\frac{\\sum_{i=1}^{n}bw_i}{n}},
+                        min \\left(
+                            bw_i,
+                            bw_i \\times \\frac{\\sum_{i=1}^{n}bw_i}{n}
+                            \\right)
+                                \\times
+                                \\frac{bw}{\\sum_{i=1}^{n}
+                                min \\left(bw_i,
+                                    bw_i \\times \\frac{\\sum_{i=1}^{n}bw_i}{n}
+                            \\right)}
+                        \\right)
+                    \\times bwdescobs_i
+
+        """
+        log.info("Calculating relays' bandwidth using Torflow method.")
+        bw_lines_tf = copy.deepcopy(bw_lines)
+        # mean (Torflow's strm_avg)
+        mu = mean([l.bw_bs_mean for l in bw_lines])
+        # filtered mean (Torflow's filt_avg)
+        muf = mean([min(l.bw_bs_mean, mu) for l in bw_lines])
+        # bw sum (Torflow's tot_net_bw or tot_sbw)
+        sum_bw = sum([l.bw_bs_mean for l in bw_lines])
+        # Torflow's clipping, not being applied
+        # hlimit = sum_bw * TORFLOW_BW_MARGIN
+        log.debug('sum %s', sum_bw)
+        log.debug('mu %s', mu)
+        log.debug('muf %s', muf)
+        # log.debug('hlimit %s', hlimit)
+        for l in bw_lines_tf:
+            if desc_obs_bws == TORFLOW_OBS_LAST:
+                desc_obs_bw = l.desc_obs_bw_bs_last
+            elif desc_obs_bws == TORFLOW_OBS_MEAN:
+                desc_obs_bw = l.desc_obs_bw_bs_mean
+            # just applying the formula above:
+            l.bw = max(round(
+                max(
+                    # ratio
+                    l.bw_bs_mean / mu,
+                    # ratio filtered
+                    min(l.bw_bs_mean, mu) / muf
+                    ) * desc_obs_bw
+                / 1000), 1)
+        return sorted(bw_lines_tf, key=lambda x: x.bw, reverse=reverse)
 
     @property
     def sum_bw(self):